Fix default value for shiftSalary. Problems with navigate by url. Переход по некоторым url не возможен при закрытой смене. Значение по умолчанию поля зарплата у босса

+++ src/main/java/com/cafe/crm/services/interfaces/shift/ShiftService.java
@@ -14,7 +14,7 @@ public interface ShiftService {
 
 	void saveAndFlush(Shift shift);
 
-	Shift newShift(int[] box, Double cashBox, Double bankCashBox);
+	Shift newShift(long[] box, Double cashBox, Double bankCashBox);
 
 	Shift findOne(Long L);
